+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/_shared/http_challenge.py
@@ -40,6 +40,10 @@ class HttpChallenge(object):
         if "authorization" not in self._parameters and "authorization_uri" not in self._parameters:
             raise ValueError("Invalid challenge parameters")
 
+        authorization_uri = self.get_authorization_server()
+        # the authoritzation server URI should look something like https://login.windows.net/tenant-id
+        self.tenant_id = authorization_uri.split("/")[-1] or None
+
         # if the response headers were supplied
         if response_headers:
             # get the message signing key and message key encryption key from the headers
@@ -81,12 +85,6 @@ class HttpChallenge(object):
     def get_scope(self):
         """ Returns the scope if present, otherwise empty string. """
         return self.get_value("scope") or ""
-
-    def get_tenant_id(self):
-        """ Returns the tenant ID parsed from the authorization server """
-        authorization_uri = self.get_authorization_server()
-        # the authoritzation server URI should look something like https://login.windows.net/tenant-id
-        return authorization_uri.split("/")[-1]
 
     def supports_pop(self):
         """ Returns True if challenge supports pop token auth else False """