+++ sdk/keyvault/azure-keyvault-certificates/azure/keyvault/certificates/_shared/challenge_auth_policy.py
@@ -42,12 +42,6 @@ def _enforce_tls(request):
         )
 
 
-def _get_challenge_scope(challenge):
-    # type: (HttpChallenge) -> str
-    # azure-identity credentials require an AADv2 scope but the challenge may specify an AADv1 resource
-    return challenge.get_scope() or challenge.get_resource() + "/.default"
-
-
 def _update_challenge(request, challenger):
     # type: (PipelineRequest, PipelineResponse) -> HttpChallenge
     """parse challenge from challenger, cache it, return it"""
@@ -77,7 +71,8 @@ class ChallengeAuthPolicy(BearerTokenCredentialPolicy):
         if challenge:
             # Note that if the vault has moved to a new tenant since our last request for it, this request will fail.
             if self._need_new_token:
-                scope = _get_challenge_scope(challenge)
+                # azure-identity credentials require an AADv2 scope but the challenge may specify an AADv1 resource
+                scope = challenge.get_scope() or challenge.get_resource() + "/.default"
                 self._token = self._credential.get_token(scope, tenant_id=challenge.tenant_id)
 
             # ignore mypy's warning -- although self._token is Optional, get_token raises when it fails to get a token
@@ -97,7 +92,8 @@ class ChallengeAuthPolicy(BearerTokenCredentialPolicy):
         # type: (PipelineRequest, PipelineResponse) -> bool
         try:
             challenge = _update_challenge(request, response)
-            scope = _get_challenge_scope(challenge)
+            # azure-identity credentials require an AADv2 scope but the challenge may specify an AADv1 resource
+            scope = challenge.get_scope() or challenge.get_resource() + "/.default"
         except ValueError:
             return False