+++ sdk/keyvault/azure-keyvault-secrets/CHANGELOG.md
@@ -3,9 +3,8 @@
 ## 4.4.0b2 (Unreleased)
 
 ### Features Added
-
-- Added support for multi-tenant authentication against Key Vault and Managed HSM when using
-  `azure-identity` 1.7.0 or newer
+- Added support for multi-tenant authentication against Managed HSM when using
+  `azure-identity` 1.7.1 or newer
   ([#20698](https://github.com/Azure/azure-sdk-for-python/issues/20698))
 
 ### Breaking Changes
@@ -13,6 +12,7 @@
 ### Bugs Fixed
 
 ### Other Changes
+- Updated minimum `azure-core` version to 1.15.0
 
 ## 4.4.0b1 (2021-09-09)